# movie-microservice/models/token.go
/*
 * @File: models.token.go
 * @Description: Defines Token information will be returned to the clients
 * @Author: Nguyen Truong Duong (seedotech@gmail.com)
 */
package models

// Token string
type Token struct {
	Token string `json:"token" example:"5bb3695b82ebac0f76e1cafa"`
}